Cart book

+++ src/app/shop/cart/cart.component.ts
@@ -20,17 +20,18 @@ export class CartComponent implements OnInit {
   constructor(private productsService: ProductsService) {}
 
   ngOnInit(): void {
+    this.productsService.getImages().subscribe((data) => this.images = data);
     this.productsService.getProducts()
       .subscribe((data) => this.updateCart(data));
   }
-
+  
   updateCart(data: Product[]) {
     this.products = data
     const newData:ProductBought[] = [];
     Object.keys(this.productsService.boughtObj).forEach(id => {
       newData.push({
         id: Number(id),
-        image: "por ahora",
+        image: "",
         name: this.products[Number(id)-1].name,
         quantity: this.productsService.boughtObj[Number(id)],
         price: this.products[Number(id)-1].price
@@ -40,20 +41,28 @@ export class CartComponent implements OnInit {
   }
 
   products: Product[] = [];
+  images: [] = [];
 
   displayedColumns: string[] = ['image', 'name', 'quantity', 'price', 'actions'];
   dataSource: ProductBought[] = [];
 
 
-  book(index: number): void {
-    console.log(index);
-    this.productsService.addItem(index);
-    this.updateCart(this.products);
+  addItem(index: number): void {
+    this.productsService.bookItem(index, 'book').subscribe((message) => {
+      message === 'Stockout' ? alert('Item out of stock') : void(0);
+      if (message === 'Booked'){
+        this.productsService.addItem(index);
+        this.updateCart(this.products);
+      }
+    });
   }
   
-  unbook(index: number): void {    
-    this.productsService.deleteItem(index);
-    this.updateCart(this.products);
+  deleteItem(index: number): void {    
+    this.productsService.bookItem(index, 'unbook').subscribe((message) => {
+      if (message !== 'Unbooked') return;
+      this.productsService.deleteItem(index);
+      this.updateCart(this.products);
+    })
   }
 
   getTotalCost(): number {
@@ -62,12 +71,12 @@ export class CartComponent implements OnInit {
     return cost;
   }
 
-  checkout(): void {
-
-  }
-
   showProducts() {
     console.log(this.dataSource);
+  }
+
+  getImage(id:number):string{
+    return this.images[id][0];
   }
 
 }
